Move all functions from utils.py to dashboard.py, they are used only there so no need to be on a separate module.
